+++ crates/rw-site/src/lib.rs
@@ -30,6 +30,7 @@
 
 pub(crate) mod site;
 pub(crate) mod site_state;
+mod page_renderer;
 mod typed_page_registry;
 
 pub use site::{PageRenderResult, RenderError, Site, SiteConfig};